Increased Spawn Delay
Need to figure out what exactly we want to do in terms of spawning Lil Guys. Right now, we just have them spawning until we reach a maximum, then respawning as needed to keep that maximum in the world, which gives no real flow of new Lil Guys. Going to ask desgin some stuff to help with this.